# Fix dirty tracking after rollback.