--- a/activerecord/lib/active_record/attribute_methods/read.rb
+++ b/activerecord/lib/active_record/attribute_methods/read.rb
@@ -9,14 +9,11 @@ module ActiveRecord
         private
 
           def define_method_attribute(name)
-            sync_with_transaction_state = "sync_with_transaction_state" if name == primary_key
-
             ActiveModel::AttributeMethods::AttrNames.define_attribute_accessor_method(
               generated_attribute_methods, name
             ) do |temp_method_name, attr_name_expr|
               generated_attribute_methods.module_eval <<-RUBY, __FILE__, __LINE__ + 1
                 def #{temp_method_name}
-                  #{sync_with_transaction_state}
                   name = #{attr_name_expr}
                   _read_attribute(name) { |n| missing_attribute(n, caller) }
                 end
@@ -36,13 +33,13 @@ module ActiveRecord
 
         primary_key = self.class.primary_key
         name = primary_key if name == "id" && primary_key
-        sync_with_transaction_state if name == primary_key
         _read_attribute(name, &block)
       end
 
       # This method exists to avoid the expensive primary_key check internally, without
       # breaking compatibility with the read_attribute API
       def _read_attribute(attr_name, &block) # :nodoc
+        sync_with_transaction_state
         @attributes.fetch_value(attr_name.to_s, &block)
       end