--- a/activerecord/lib/active_record/attribute_methods/write.rb
+++ b/activerecord/lib/active_record/attribute_methods/write.rb
@@ -13,15 +13,12 @@ module ActiveRecord
         private
 
           def define_method_attribute=(name)
-            sync_with_transaction_state = "sync_with_transaction_state" if name == primary_key
-
             ActiveModel::AttributeMethods::AttrNames.define_attribute_accessor_method(
               generated_attribute_methods, name, writer: true,
             ) do |temp_method_name, attr_name_expr|
               generated_attribute_methods.module_eval <<-RUBY, __FILE__, __LINE__ + 1
                 def #{temp_method_name}(value)
                   name = #{attr_name_expr}
-                  #{sync_with_transaction_state}
                   _write_attribute(name, value)
                 end
               RUBY
@@ -40,21 +37,21 @@ module ActiveRecord
 
         primary_key = self.class.primary_key
         name = primary_key if name == "id" && primary_key
-        sync_with_transaction_state if name == primary_key
         _write_attribute(name, value)
       end
 
       # This method exists to avoid the expensive primary_key check internally, without
       # breaking compatibility with the write_attribute API
       def _write_attribute(attr_name, value) # :nodoc:
+        sync_with_transaction_state
         @attributes.write_from_user(attr_name.to_s, value)
         value
       end
 
       private
         def write_attribute_without_type_cast(attr_name, value)
-          name = attr_name.to_s
-          @attributes.write_cast_value(name, value)
+          sync_with_transaction_state
+          @attributes.write_cast_value(attr_name.to_s, value)
           value
         end