@kaspth 2019-04-16
Nice! Is there any performance impact of running `sync_with_transaction_state` more often now?