@kamipo 2019-04-16
https://gist.github.com/kamipo/cb6138e01f2f706c4f2e87a1282c8428

Before (d4e2824d8d9aea6d95e78275c0107bbbc48300d9 before #35933):

```
Warming up --------------------------------------
             user.id    82.135k i/100ms
          user['id']    63.284k i/100ms
           user.name   154.252k i/100ms
        user['name']    92.613k i/100ms
       user.changed?    10.552k i/100ms
 user.saved_changes?    25.754k i/100ms
Calculating -------------------------------------
             user.id      1.101M (±10.9%) i/s -      5.421M in   5.045868s
          user['id']    794.397k (± 1.0%) i/s -      3.987M in   5.019332s
           user.name      2.637M (± 2.5%) i/s -     13.266M in   5.033214s
        user['name']      1.276M (± 1.5%) i/s -      6.390M in   5.009238s
       user.changed?    111.988k (± 9.7%) i/s -    559.256k in   5.052791s
 user.saved_changes?    278.738k (±12.3%) i/s -      1.365M in   5.018580s
```

After (this PR 63ff495bdf90e0ab20114a49db5cffe3cb9ef2fd):

```
Warming up --------------------------------------
             user.id   118.536k i/100ms
          user['id']    94.816k i/100ms
           user.name   151.977k i/100ms
        user['name']    92.442k i/100ms
       user.changed?    18.335k i/100ms
 user.saved_changes?    24.328k i/100ms
Calculating -------------------------------------
             user.id      2.108M (± 1.7%) i/s -     10.550M in   5.007100s
          user['id']      1.294M (± 1.1%) i/s -      6.542M in   5.057655s
           user.name      2.544M (± 1.0%) i/s -     12.766M in   5.019074s
        user['name']      1.290M (± 1.0%) i/s -      6.471M in   5.018621s
       user.changed?    191.417k (±11.6%) i/s -    953.420k in   5.097952s
 user.saved_changes?    263.707k (±12.1%) i/s -      1.289M in   5.012170s
```

`user.id` is 2x faster by 8180c3961063c5786305a94a7d204ab4bc72c381.
Perhaps `user['name']` is a bit faster and `user.name` is a bit slower, but not significant.