@shioyama 2019-04-21
Sorry, I see now that @kamipo's comment mentions #35933. But the point remains then that `user.id` is not improved significantly by this change, whereas `user.name` performance drops, I think significantly but that needs to be confirmed.