@shioyama 2019-04-21
The commit that gives (most of?) the 2x boost is I believe 8180c3961063c5786305a94a7d204ab4bc72c381.

I might be totally not understanding something here, if so I apologize.